pyproject.toml
[project]
name = "discord-stats"
version = "0.1.0"
description = "Discord bot for fetching and reporting server statistics"
authors = [
  { name = "Martin Meneval", email = "34133448+martinmeneval@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "discord-py (>=2.5.2,<3.0.0)",
  "pydantic (>=2.11.0,<3.0.0)",
  "python-dateutil (>=2.9.0.post0,<3.0.0)",
  "click (>=8.1.7,<9.0.0)",
]

[project.scripts]
discord-stats = "discord_stats:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
